requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
chardet==4.0.0
colorzero==1.1
discord.py==1.6.0
gpiozero==1.5.1
idna==3.1
multidict==5.1.0
RPi.GPIO==0.7.0
typing-extensions==3.7.4.3
yarl==1.6.3
